+++ src/components/screens/quizzes/ScreenAddQuiz.tsx
@@ -11,9 +11,9 @@ import { createAlert } from "../../../helperScripts/createAlert";
 import i18n from "../../../locales";
 import { RequestFactory } from "../../../api/requests/RequestFactory";
 import { EndpointsChapter } from "../../../api/endpoints/EndpointsChapter";
-import { ScreenCourseTabsRouteProp } from "./ScreenCourseTabs";
-import { ScreenCourseOverviewNavigationProp } from "./ScreenCourseOverview";
-import { QuestionCard } from "../../QuestionCard";
+import { ScreenCourseTabsRouteProp } from "../course/ScreenCourseTabs";
+import { ScreenCourseOverviewNavigationProp } from "../course/ScreenCourseOverview";
+import { QuestionCard } from "../../cards/QuestionCard";
 import { ScrollView } from "react-native-gesture-handler";
 import { IQuestionMultipleChoice, IQuestionNumeric, IQuestionSingleChoice } from "../../../types/IQuestion";
 import { IUser } from "../../../types/IUser";
